requirements.txt
Flask
Flask-SocketIO
simple-websocket
python-dotenv
Werkzeug
gunicorn
gevent
gevent-websocket
requests
cryptography
